backend/requirements.txt
fastapi
uvicorn[standard]
openai
speechrecognition
pyttsx3
python-dotenv
pyautogui
psutil
websockets
sounddevice
faster-whisper